moved interactions functions to controller and put each into an activate function called by activate listeners
